Modify the geotrellis.spark.render package to be more in line with design patterns
